sqlite, some bugs fixed
